Moved components into a components folder
