+++ frontend/README.md
@@ -47,11 +47,11 @@ These are links to some of the files that we have implemented/modified when deve
   Includes all the necessary modules and components
 - **AppComponent** | [TS](./src/app/app.component.ts), [HTML](./src/app/app.component.html)  
   Main component of the frontend application
-- **TodoListComponent** | [TS](./src/app/todo-list/todo-list.component.ts), [HTML](./src/app/todo-list/todo-list.component.html)  
+- **TodoListComponent** | [TS](src/app/components/todo-list/todo-list.component.ts), [HTML](src/app/components/todo-list/todo-list.component.html)  
   Represents a todo-list
-- **TodoItemComponent** | [TS](./src/app/todo-list/todo-item.component.ts), [HTML](./src/app/todo-list/todo-item.component.html)  
+- **TodoItemComponent** | [TS](src/app/components/todo-list/todo-item.component.ts), [HTML](src/app/components/todo-list/todo-item.component.html)  
   Represents a todo-item
-- **UserLoginComponent** | [TS](./src/app/user-login/user-login.component.ts), [HTML](./src/app/user-login/user-login.component.html)  
+- **UserLoginComponent** | [TS](src/app/components/user-login/user-login.component.ts), [HTML](src/app/components/user-login/user-login.component.html)  
   Represents the user login section
 - **[AuthInterceptor](./src/app/auth/auth.interceptor.ts)**  
   Intercepts all HTTP requests to add the user token if a user is logged in.
