+++ frontend/src/app/components/home/home.component.html
@@ -1,0 +1,1 @@
+<h1>Welcome</h1>
