+++ frontend/src/app/components/todo-list/todo-item/todo-item.component.ts
@@ -1,5 +1,5 @@
 import { Component, EventEmitter, Input, Output } from '@angular/core';
-import { TodoItem } from '../../models/todo-item.model';
+import { TodoItem } from '../../../models/todo-item.model';
 
 @Component({
   selector: 'app-todo-item',
